# reviews-service/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Swagger documentation",
    "version": "1.0.0"
  },
  "host": "localhost:3002",
  "basePath": "/",
  "tags": [
    {
      "name": "Reviews",
      "description": "Reviews related endpoints"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "description": "",
        "parameters": [
          {
            "name": "movieId",
            "in": "query",
            "description": "ID of the movie to filter reviews",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "List of reviews",
            "schema": {
              "$ref": "#/definitions/GetReview"
            }
          },
          "400": {
            "description": "Invalid movieId parameter",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "movieId must be a number"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Access denied. Token missing.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access denied. Token missing."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Access forbidden: Invalid token.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access forbidden: Invalid token."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reviews"
        ],
        "description": "Create a new review",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "review",
            "in": "body",
            "description": "Review object to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateReview"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Review created successfully",
            "schema": {
              "$ref": "#/definitions/CreateReview"
            }
          },
          "401": {
            "description": "Access denied. Token missing.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access denied. Token missing"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Access forbidden: Invalid token.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access forbidden: Invalid token."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Reviews"
        ],
        "description": "Delete all reviews of a given movie by movieId",
        "parameters": [
          {
            "name": "movieId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "movieId",
            "in": "path",
            "description": "ID of the movie whose reviews should be deleted",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "All reviews for the movie deleted successfully."
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "No reviews found for the given movieId.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No reviews found for this movie"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/reviews/{id}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "description": "Get a review by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the review to retrieve"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Review found",
            "schema": {
              "$ref": "#/definitions/GetReview"
            }
          },
          "401": {
            "description": "Access denied. Token missing.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access denied. Token missing."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Access forbidden: Invalid token.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access forbidden: Invalid token."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Review not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Review not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Reviews"
        ],
        "description": "Delete a review by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the review to delete"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Review deleted successfully"
          },
          "401": {
            "description": "Access denied. Token missing.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access denied. Token missing."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Access forbidden: Invalid token.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access forbidden: Invalid token."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Review not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Review not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Reviews"
        ],
        "description": "Update a review by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the review to update"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Review data to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateReview"
            }
          },
          {
            "name": "review",
            "in": "body",
            "description": "Updated review object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetReview"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review updated successfully",
            "schema": {
              "$ref": "#/definitions/GetReview"
            }
          },
          "401": {
            "description": "Access denied. Token missing.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access denied. Token missing."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Access forbidden: Invalid token.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access forbidden: Invalid token."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Review not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Review not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "GetReview": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "movieId": {
          "type": "number",
          "example": 1
        },
        "userId": {
          "type": "number",
          "example": 1
        },
        "text": {
          "type": "string",
          "example": "Great movie!"
        }
      }
    },
    "CreateReview": {
      "type": "object",
      "properties": {
        "movieId": {
          "type": "number",
          "example": 1
        },
        "userId": {
          "type": "number",
          "example": 1
        },
        "text": {
          "type": "string",
          "example": "Great movie!"
        }
      }
    }
  }
}